+++ src/Mscc.GenerativeAI.Web/CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## 1.1.0
 
+### Added
+
+- implement generic access to GenerativeModel
+
 ### Changed
 
 - bump version